merged: changes from linuxport r20803-21500, 21500-22000
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/unified-pvr@22454 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/lib/cximage-6.0/CxImage/ximatran.cpp
@@ -1111,7 +1111,7 @@ bool CxImage::Resample(long newx, long newy, int mode, CxImage* iDst)
 					naTemp = naCarry;
 					naCarry = naAccu;
 					naAccu = naTemp;
-					memset(naCarry, 0, sizeof(int) * 3);    // need only to set first pixel zero
+					memset(naCarry, 0, sizeof(long) * 3);    // need only to set first pixel zero
 					pDest = newImage.info.pImage + (++v * newImage.info.dwEffWidth);
 					fEndY += yScale;
 				}
